Restructure backend package structure
