+++ backend/__test__/postgres/user-query.test.js
@@ -3,9 +3,8 @@
 
 'use strict';
 
-const userQuery = require('../postgres/user-query.js');
-// const recipeQuery = require('../service/recipe-query.js');
-// const userQuery = require('../service/user-query.js');
+const userQuery = require('../../src/postgres/user-query.js');
+
 jest.mock('../model/database.js', () => {
   const SequelizeMock = require('sequelize-mock');
   return { sequelize: new SequelizeMock() };
@@ -16,8 +15,6 @@ jest.mock('../model/User.js', () => () => {
   const dbMock = new SequelizeMock();
   return dbMock.define('User', {
     uuid: '9f50a9ff-273b-42df-8438-9e5adb6c675e',
-  }, {
-    
   });
 });
